backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
mysql-connector-python==9.1.0
gunicorn==23.0.0
openai==1.58.1
pillow==11.1.0
moviepy==2.1.2
werkzeug==3.1.3
bcrypt==4.2.1
python-dotenv==1.0.1
pyjwt==2.10.1
pandas==2.2.3
numpy==2.2.1
scikit-learn==1.6.1
faiss-cpu==1.9.0.post1
requests==2.32.3
